send-document-mcp/pyproject.toml
[project]
name = "send-document-mcp"
version = "0.1.0"
description = "MCP server for sending agent documents to leads"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
]



